# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: study_vector API
  version: 0.2.0
  summary: Milvus 向量数据库实战教学平台 — 统一 REST API
  description: '# study_vector API 契约


    ## 教学目标

    1. **多向量库适配**：通过 `VectorRepository` 抽象支持多种向量库（V1：Milvus）

    2. **教学友好**：每个字段都有 description 和 examples，可直接作为教学示例

    3. **契约稳定**：本文件是 API 的唯一事实源，后端实现 / 前端类型 / 契约测试都从这里派生


    ## 通用约定

    - 所有响应统一格式：`{ code, message, data }`

    - `code = 0` 表示业务成功；非 0 是错误码（与 HTTP 状态码正交）

    - 业务异常用 `4xx`，后端错误用 `5xx`，向量库后端错误统一 `502`

    - 业务 ID 字段统一用字符串（即便 Milvus 内部可能是 INT64）；向量库自动转换


    ## 路径前缀

    - 所有 API 前缀 `/api/v1`


    ## 后端实现优先级

    | 阶段 | 端点 | 状态 |

    |---|---|---|

    | 1.1 | health / collections / vectors | ✅ 已实现 |

    | 1.3 | collections/{name}/load, /release | 🔜 |

    | 1.4 | collections/{name}/indexes | 🔜 |

    | 1.5 | collections/{name}/partitions | 🔜 |

    | 1.6 | collections/{name}/alias | 🔜 |

    | 1.7 | collections/{name}/query | 🔜 |

    | 1.8 | collections/{name}/hybrid_search | 🔜 |

    | 1.9 | databases / users | 🔜 |

    '
  contact:
    name: study_vector contributors
  license:
    name: MIT
servers:
- url: http://localhost:8000
  description: 本地开发（Python FastAPI）
...
paths:
...
  /api/v1/collections:
    get:
      tags:
      - collections
      summary: 列出所有集合
      description: '返回当前 DB 下所有集合的名字。


        教学点：Milvus 用 etcd 存元数据；列集合是元数据查询，不会扫数据。

        '
      operationId: listCollections
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        type: string
                      example:
                      - demo
                      - tutorial_index
    post:
      tags:
      - collections
      summary: 创建集合
      description: '根据 schema 创建一个空集合，并自动建好 AUTOINDEX + 加载到内存。


        教学点：

        - 集合 = 表 + schema + 索引

        - 索引建好才能检索

        - **AUTOINDEX** 是 Milvus 的"自动选最优索引"机制（生产推荐）

        '
      operationId: createCollection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionSchema'
              additionalProperties: false
            examples:
              tiny:
                summary: 最小可用（教学示例用 4 维）
                value:
                  name: demo
                  dimension: 4
                  metric: COSINE
              openai_1536:
                summary: OpenAI text-embedding-3-small
                value:
                  name: openai_chunks
                  dimension: 1536
                  metric: COSINE
                  description: 文档块向量库
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        name:
                          type: string
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BackendError'
        '400': &id001
          $ref: '#/components/responses/BadRequestBody'
  /api/v1/collections/{name}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
      description: 集合名
    get:
      tags:
      - collections
      summary: 查看集合详情
      description: '返回集合的完整元信息：维度 / 度量 / 行数 / 加载状态 / 索引。

        '
      operationId: getCollection
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/CollectionInfo'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - collections
      summary: 删除集合
      description: '删除集合及其所有数据。**不可逆**。

        生产建议：删除前先 release + 备份元数据。

        '
      operationId: dropCollection
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      properties:
                        name:
                          type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/{name}/load:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - collections
      summary: 加载到内存
      description: '把集合索引加载到查询节点内存中。

        **检索前必须先 load**。


        教学点：

        - 大集合 load 耗时长（按数据量级分钟级）

        - `replica_number` 控制副本数（提升读 QPS）

        '
      operationId: loadCollection
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadRequest'
              additionalProperties: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': &id002
          $ref: '#/components/responses/NotFound'
        '422': &id003
          $ref: '#/components/responses/BadRequest'
        '400': *id001
  /api/v1/collections/{name}/release:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - collections
      summary: 从内存释放
      description: '从查询节点卸载集合。**不影响数据**，只释放内存。

        '
      operationId: releaseCollection
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
        '422':
          $ref: '#/components/responses/BadRequest'
        '400': *id001
...
  /api/v1/collections/{name}/indexes:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - indexes
      summary: 列出索引
      operationId: listIndexes
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        $ref: '#/components/schemas/IndexInfo'
        '404': *id002
    post:
      tags:
      - indexes
      summary: 建索引
      description: '为向量字段建索引。


        教学点：

        - 集合创建时默认建 AUTOINDEX；如要换 IVF / HNSW 需先 drop 再建

        - 建索引是**阻塞**操作（大集合可能数十分钟）

        - 索引建好才能 release；建议建好后再 release 释放构建期内存

        '
      operationId: createIndex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIndexRequest'
              # 显式不开 additionalProperties: false，避免 schemathesis 4.x strict-mode
              # 在 CreateIndexRequest 内含 free-form `params` 对象时触发 Unsatisfiable
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
        '422': *id003
        '400': *id001
  /api/v1/collections/{name}/indexes/{field_name}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    - name: field_name
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - indexes
      summary: 删除索引
      operationId: dropIndex
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
...
  /api/v1/collections/{name}/alias:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - alias
      summary: 给集合绑定 alias
      description: '一个集合可以有多个 alias；一个 alias 只能指向一个集合。

        用于零停机切换（reindex 流程：建新集合 → 写双写 → alias 切换 → 删旧集合）。

        '
      operationId: createAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - alias
              properties:
                alias:
                  type: string
                  minLength: 1
                  maxLength: 255
                  pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
                  description: alias 名
              additionalProperties: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
        '422': *id003
        '400': *id001
    get:
      tags:
      - alias
      summary: 列出集合的 alias
      operationId: listAliases
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        type: string
        '404': *id002
  /api/v1/collections/{name}/alias/{alias}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    - name: alias
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - alias
      summary: 解绑 alias
      operationId: dropAlias
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
  /api/v1/databases:
    get:
      tags:
      - databases
      summary: 列出数据库
      operationId: listDatabases
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    data:
                      type: array
                      items:
                        type: string
    post:
      tags:
      - databases
      summary: 创建数据库
      operationId: createDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
                  description: 数据库名
              additionalProperties: false
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '422': *id003
        '400': *id001
  /api/v1/databases/{name}:
    parameters:
    - name: name
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - databases
      summary: 删除数据库
      operationId: dropDatabase
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': *id002
...
components:
  schemas:
    ApiResponse:
      type: object
      description: '**统一响应格式**：所有业务接口都包成这个形状。

        - `code = 0` 表示成功

        - `code ≠ 0` 表示错误（业务 / 校验 / 后端），对应 HTTP 状态码见 error codes 表

        '
      required:
      - code
      - message
      - data
      properties:
        code:
          type: integer
          description: 业务码。0=成功；非 0=错误（见错误码表）
          example: 0
        message:
          type: string
          description: 人类可读消息
          example: success
        data:
          description: 业务负载；类型由具体端点决定
          oneOf:
          - type: object
          - type: array
          - type: string
          - type: number
          - type: boolean
          - type: 'null'
      additionalProperties: false
...
    DistanceMetric:
      type: string
      enum:
      - COSINE
      - L2
      - IP
      - HAMMING
      - JACCARD
      description: '相似度度量。

        - **COSINE**（余弦）：只看方向不看模长，文本嵌入**最常用**；归一化向量等价于内积

        - **L2**（欧氏距离）：两点直线距离；对向量模长敏感

        - **IP**（内积）：未归一化向量的"对齐度"

        - **HAMMING / JACCARD**：二值向量专用

        '
      example: COSINE
    IndexType:
      type: string
      enum:
      - FLAT
      - IVF_FLAT
      - IVF_SQ8
      - IVF_PQ
      - HNSW
      - ANNOY
      - AUTOINDEX
      - DISKANN
      description: '索引类型。

        - **AUTOINDEX**（生产推荐）：Milvus 自动选最优

        - **HNSW**：高召回 / 内存允许

        - **IVF_FLAT / SQ8 / PQ**：大数据 / 不同压缩比

        - **FLAT**：精确 / 小数据

        - **ANNOY**：只读场景

        - **DISKANN**：内存不够 / 大规模

        '
      example: AUTOINDEX
    VectorType:
      type: string
      enum:
      - FloatVector
      - BinaryVector
      - Float16Vector
      - BFloat16Vector
      - SparseFloatVector
      - Int8Vector
      description: '向量数据类型。

        - **FloatVector**（默认）：通用 embedding，4×D bytes

        - **BinaryVector**：二值特征，D/8 bytes；维度必须是 8 的倍数；配 HAMMING/JACCARD

        - **Float16Vector**：半精度，2×D bytes；配 COSINE/L2/IP

        - **BFloat16Vector**：训练用，2×D bytes；配 COSINE/L2/IP

        - **SparseFloatVector**：稀疏关键词，IP 度量；向量以 dict 形式传入

        - **Int8Vector**：量化，D bytes；配 L2

        '
      example: FloatVector
    ConsistencyLevel:
      type: string
      enum:
      - Strong
      - Session
      - Bounded
      - Eventually
      description: '一致性等级。

        - **Strong**：强一致（最高延迟）

        - **Session**：单会话内一致（默认）

        - **Bounded**：可容忍旧数据

        - **Eventually**：最终一致（最快）

        '
      example: Session
    CollectionSchema:
      type: object
      description: '业务层集合 schema。

        统一约定：所有集合自动包含三字段

        - `<primary_field>` (VARCHAR, 主键)

        - `<vector_field>` (FLOAT_VECTOR, 维度=dimension)

        - `payload` (JSON, 业务元数据，可检索)

        '
      required:
      - name
      - dimension
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
          description: '集合名。命名规则：

            - 以字母开头

            - 后续字符：字母 / 数字 / 下划线

            - 长度 1-255

            '
          example: demo_collection
        dimension:
          type: integer
          minimum: 1
          maximum: 65536
          description: '向量维度。

            - 与使用的嵌入模型输出一致（如 OpenAI text-embedding-3-small=1536；BGE-base=768）

            - **一旦创建不可修改**

            '
          example: 1536
        metric:
          $ref: '#/components/schemas/DistanceMetric'
          description: 距离/相似度度量。
          example: COSINE
        vector_type:
          $ref: '#/components/schemas/VectorType'
          description: '向量数据类型。默认 `FloatVector`。

            - `BinaryVector` 维度必须是 8 的倍数（按位打包）

            - `SparseFloatVector` 的 `dimension` 表示 NNZ 上限，向量以 `dict {idx: value}` 形式存储

            - 选错类型 = 维度/度量/索引/距离计算全错

            '
          example: FloatVector
        primary_field:
          type: string
          default: id
          description: 主键字段名。V1 固定 VARCHAR。
          example: id
        vector_field:
          type: string
          default: vector
          description: 向量字段名。
          example: vector
        description:
          type: string
          nullable: true
          description: 集合描述（人类可读）。
          example: Demo collection for tutorial
        index_type:
          $ref: '#/components/schemas/IndexType'
          description: 索引类型。V1 推荐 `AUTOINDEX`（Milvus 自适应）。
          example: AUTOINDEX
        consistency_level:
          $ref: '#/components/schemas/ConsistencyLevel'
          description: 一致性等级。CAP 三角取舍：Strong 慢但准；Eventually 快但可能查到旧数据。
          example: Session
      additionalProperties: false
      # 跨字段约束：vector_type ↔ metric 兼容性
      # （OpenAPI 3.1 = JSON Schema 2020-12，支持 if/then/else）
      allOf:
        - if:
            properties:
              vector_type:
                const: BinaryVector
            required:
              - vector_type
          then:
            properties:
              metric:
                enum:
                  - HAMMING
                  - JACCARD
              dimension:
                multipleOf: 8
        - if:
            properties:
              vector_type:
                const: Int8Vector
            required:
              - vector_type
          then:
            properties:
              metric:
                enum:
                  - L2
        - if:
            properties:
              vector_type:
                const: SparseFloatVector
            required:
              - vector_type
          then:
            properties:
              metric:
                enum:
                  - IP
        - if:
            properties:
              vector_type:
                enum:
                  - FloatVector
                  - Float16Vector
                  - BFloat16Vector
            required:
              - vector_type
          then:
            properties:
              metric:
                enum:
                  - COSINE
                  - L2
                  - IP
    CollectionInfo:
      type: object
      description: 集合运行时信息
      required:
      - name
      - dimension
      - metric
      - row_count
      - loaded
      properties:
        name:
          type: string
          description: 集合名。
          example: demo
        dimension:
          type: integer
          description: 向量维度。
          example: 1536
        metric:
          $ref: '#/components/schemas/DistanceMetric'
          description: 距离/相似度度量。
        vector_type:
          $ref: '#/components/schemas/VectorType'
          description: 向量数据类型。
          example: FloatVector
        row_count:
          type: integer
          description: 当前集合内向量条数。
          example: 1000
        loaded:
          type: boolean
          description: 是否已加载到内存（只有 loaded=true 才能检索）。
          example: true
        consistency_level:
          $ref: '#/components/schemas/ConsistencyLevel'
          description: 一致性等级（创建时设置，运行时不可改）
          example: Session
        indexes:
          type: array
          description: 集合上的索引列表。
          items:
            $ref: '#/components/schemas/IndexInfo'
        created_at:
          type: string
          format: date-time
          nullable: true
          description: 集合创建时间（UTC）。
          example: '2024-01-15T08:00:00Z'
      additionalProperties: false
    LoadRequest:
      type: object
      description: 加载集合的请求体
      properties:
        replica_number:
          type: integer
          minimum: 1
          maximum: 16
          default: 1
          description: '副本数。每个副本占一份内存；提升读 QPS。

            不能超过 QueryNode 数量。

            '
        partition_names:
          type: array
          items:
            type: string
          description: 限定加载的分区（不指定则全量加载）
      additionalProperties: false
...
    IndexInfo:
      type: object
      description: 索引元信息
      required:
      - field_name
      - index_type
      - metric_type
      properties:
        field_name:
          type: string
          description: 索引对应的字段名（一般是向量字段）。
          example: vector
        index_type:
          $ref: '#/components/schemas/IndexType'
          description: 索引算法类型。
          example: HNSW
        metric_type:
          $ref: '#/components/schemas/DistanceMetric'
          description: 索引使用的距离度量。
          example: COSINE
        params:
          type: object
          additionalProperties: true
          description: '索引参数。

            - HNSW: `{"M": 16, "efConstruction": 64}`

            - IVF_PQ: `{"nlist": 1024, "m": 8, "nbits": 8}`

            '
          example:
            M: 16
            efConstruction: 64
      additionalProperties: false
    CreateIndexRequest:
      type: object
      description: 建索引请求
      required:
      - field_name
      - index_type
      additionalProperties: false
      properties:
        field_name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: ^[a-zA-Z][a-zA-Z0-9_]*$
          description: 要建索引的字段名（一般是向量字段）。
          example: vector
        index_type:
          $ref: '#/components/schemas/IndexType'
          description: 索引算法类型。
          example: HNSW
        metric_type:
          $ref: '#/components/schemas/DistanceMetric'
          description: 索引距离度量（必须与集合度量一致）。
          example: COSINE
        params:
          type: object
          # 不写 additionalProperties: true（避免与父级 additionalProperties: false
          # 在 schemathesis strict-mode 下冲突）；保持 {} 表示任意键值对。
          description: '索引参数。常用：

            - HNSW: `{"M": 16, "efConstruction": 64}`

            - IVF_FLAT: `{"nlist": 1024}`

            - IVF_PQ: `{"nlist": 1024, "m": 8, "nbits": 8}`

            '
          example:
            M: 16
            efConstruction: 64
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Conflict:
      description: 资源已存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    BadRequest:
      description: 参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    BackendError:
      description: 向量库后端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    BadRequestBody:
      description: 请求体解析失败（JSON 格式错误等）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'